+++ src/step.rs
@@ -1,5 +1,12 @@
 use ::core::convert::TryInto;
 
+pub trait Bounded {
+    /// Those two methods could also be from num_traits::Bounded
+    fn min_value() -> Self;
+    fn max_value() -> Self;
+}
+
+
 /// Similar to std::iter::Step, but the latter is unstable and cannot be used
 /// in this package.  It also doesn't provide support for starting from lowest
 /// value valid for the type for instance.
@@ -7,10 +14,6 @@ pub trait Step
 where
     Self: ::core::marker::Sized,
 {
-    /// Those two methods could also be from num_traits::Bounded
-    fn min_value() -> Self;
-    fn max_value() -> Self;
-
     fn forward(&self, step: usize) -> Option<Self>;
     fn backward(&self, step: usize) -> Option<Self>;
 
@@ -29,12 +32,6 @@ where
 macro_rules! step_for_int {
     ($t:tt) => {
         impl Step for $t {
-            fn min_value() -> Self {
-                Self::MIN
-            }
-            fn max_value() -> Self {
-                Self::MAX
-            }
             fn forward(&self, step: usize) -> Option<Self> {
                 self.checked_add(step as Self)
             }
@@ -51,6 +48,19 @@ macro_rules! step_for_int {
     };
 }
 
+macro_rules! bounded_for_type {
+    ($t:tt) => {
+        impl Bounded for $t {
+            fn min_value() -> Self {
+                Self::MIN
+            }
+            fn max_value() -> Self {
+                Self::MAX
+            }
+        }
+    }
+}
+
 step_for_int!(u8);
 step_for_int!(u16);
 step_for_int!(u32);
@@ -62,3 +72,26 @@ step_for_int!(i32);
 step_for_int!(i64);
 step_for_int!(isize);
 step_for_int!(usize);
+
+bounded_for_type!(u8);
+bounded_for_type!(u16);
+bounded_for_type!(u32);
+bounded_for_type!(u64);
+bounded_for_type!(u128);
+bounded_for_type!(i8);
+bounded_for_type!(i16);
+bounded_for_type!(i32);
+bounded_for_type!(i64);
+bounded_for_type!(isize);
+bounded_for_type!(usize);
+bounded_for_type!(f32);
+bounded_for_type!(f64);
+
+impl Bounded for char {
+    fn min_value() -> Self {
+        '\0'
+    }
+    fn max_value() -> Self {
+        Self::MAX
+    }
+}